+++ contracts/marketplace/src/testing/setup/constants.rs
@@ -1,0 +1,26 @@
+// pub const TOKEN_ID: u32 = 123;
+pub const CREATOR_INITIAL_BALANCE: u128 = 5000000000;
+// pub const LISTING_FEE: u128 = 0;
+pub const INITIAL_BALANCE: u128 = 5000000000;
+
+pub const CREATION_FEE: u128 = 5_000_000_000;
+pub const MINT_PRICE: u128 = 100_000_000;
+
+pub const MAX_TOKEN_LIMIT: u32 = 10000;
+
+pub const MIN_MINT_PRICE: u128 = 50_000_000;
+pub const AIRDROP_MINT_PRICE: u128 = 0;
+pub const MINT_FEE_FAIR_BURN: u64 = 1_000; // 10%
+pub const AIRDROP_MINT_FEE_FAIR_BURN: u64 = 10_000; // 100%
+pub const SHUFFLE_FEE: u128 = 500_000_000;
+pub const MAX_PER_ADDRESS_LIMIT: u32 = 50;
+
+pub const LISTING_FEE: u128 = 0;
+pub const SECOND_BIDDER_INITIAL_BALANCE: u128 = 2000;
+
+// Governance parameters
+pub const TRADING_FEE_BPS: u64 = 200; // 2%
+pub const MIN_EXPIRY: u64 = 24 * 60 * 60; // 24 hours (in seconds)
+pub const MAX_EXPIRY: u64 = 180 * 24 * 60 * 60; // 6 months (in seconds)
+pub const MAX_FINDERS_FEE_BPS: u64 = 1000; // 10%
+pub const BID_REMOVAL_REWARD_BPS: u64 = 500; // 5%
